Add arpcgen for making small RPC clients in C.
git-svn-id: svn://okws.org/sfslite2/trunk@2091 7d287422-6eea-da11-b677-00123f2a783a

+++ libsfs/suio.c
@@ -362,7 +362,7 @@ __suio_copy (struct suio *uio, const char *data, u_int len)
   }
 
   dp = uio->uio_dp;
-  n = len < (u_int) uio->uio_dspace ? len : uio->uio_dspace;
+  n = len < (u_int) uio->uio_dspace ? len : (u_int) uio->uio_dspace;
   memcpy (dp, data, n);
   __suio_addiov (uio, dp, n);
